pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qbot"
version = "0.1.2"
description = "A cryptocurrency quantitative analysis and trading package"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = ["crypto", "trading", "quantitative", "analysis", "bot"]
authors = [
    { name = "Jay Suh", email = "jayhansuh@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "ccxt>=3.0.0",
    "python-binance>=1.0.0",
    "ta>=0.10.0",  # Technical Analysis library
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "requests>=2.26.0",
    "python-dotenv>=0.19.0",
    "websockets>=10.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "pyarrow>=14.0.1",
    "ipykernel>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/jayhansuh/qbot-sdk"
Documentation = "https://github.com/jayhansuh/qbot-sdk#readme"
Repository = "https://github.com/jayhansuh/qbot-sdk.git"
Issues = "https://github.com/jayhansuh/qbot-sdk/issues"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov"

[tool.hatch.build]
only-packages = true
packages = ["qbot"]
